# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /top-students:
    get:
      operationId: getTopStudents
      tags: [topStudents]
      summary: List top students
      parameters:
        - name: class
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Top students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopStudent"
    post:
      operationId: createTopStudent
      tags: [topStudents]
      summary: Add top student (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopStudentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopStudent"
  /top-students/{id}:
    patch:
      operationId: updateTopStudent
      tags: [topStudents]
      summary: Update top student (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopStudentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopStudent"
    delete:
      operationId: deleteTopStudent
      tags: [topStudents]
      summary: Delete top student (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Notices ---
...
components:
  schemas:
...
    TopStudent:
      type: object
      required: [id, name, class, rank, year]
      properties:
        id:
          type: integer
        rank:
          type: integer
        name:
          type: string
        fatherName:
          type: ["string", "null"]
        class:
          type: string
        percentage:
          type: ["number", "null"]
        totalMarks:
          type: ["integer", "null"]
        grade:
          type: ["string", "null"]
        year:
          type: string
        photoUrl:
          type: ["string", "null"]
    TopStudentInput:
      type: object
      required: [name, class, rank, year]
      properties:
        rank:
          type: integer
        name:
          type: string
        fatherName:
          type: string
        class:
          type: string
        percentage:
          type: number
        totalMarks:
          type: integer
        grade:
          type: string
        year:
          type: string
        photoUrl:
          type: string
    TopStudentUpdate:
      type: object
      properties:
        rank:
          type: integer
        name:
          type: string
        fatherName:
          type: string
        class:
          type: string
        percentage:
          type: number
        totalMarks:
          type: integer
        grade:
          type: string
        year:
          type: string
        photoUrl:
          type: string

    # Notices
...